+++ Services/Administration/OJS.Services.Administration.Business/Submissions/SubmissionsBusinessService.cs
@@ -90,7 +90,7 @@ namespace OJS.Services.Administration.Business.Submissions
                         s.Id,
                         s.ParticipantId,
                         CorrectTestRuns = s.TestRuns.Count(t =>
-                            t.ResultType == TestRunResultType.CorrectAnswer &&
+                            t.ResultType == (int)TestRunResultType.CorrectAnswer &&
                             !t.Test.IsTrialTest),
                         AllTestRuns = s.TestRuns.Count(t => !t.Test.IsTrialTest),
                         MaxPoints = s.Problem!.MaximumPoints,
